Move file IO to read.{cpp,h}
